Fix waiting on reachability of SSH port of the GenericLinuxHost
Remove retry count handling from the 'get' method because restart is not
supported and retry makes no sense without it. There was also a bug --
when machine was not reachable after timeout, exception was written and
the code continued wihtout any failure. Now it fails after the timeout
when fail_ctrl has been set to 'ignore' or exits if fail_ctrl set
otherwise.

Signed-off-by: Martin Banszel <martinx.banszel@intel.com>